requirements.txt
pandas==1.1.3
requests==2.24.0
Pillow
